+++ src/demo/hud.c
@@ -392,13 +392,13 @@ hud_print_finished(elem* list){
       if(ncplane_printf_yx(hud, line, 1, "%d", e->frames) < 0){
         return -1;
       }
-      char buf[PREFIXCOLUMNS];
-      ncmetric(e->totalns, NANOSECS_IN_SEC, buf, 0, 1000, '\0');
-      for(size_t x = 6 ; x < 14 - strlen(buf) ; ++x){
+      char buf[PREFIXCOLUMNS + 2];
+      ncnmetric(e->totalns, sizeof(buf), NANOSECS_IN_SEC, buf, 0, 1000, '\0');
+      for(int x = 6 ; x < 14 - ncstrwidth(buf) ; ++x){
         nccell ci = CELL_TRIVIAL_INITIALIZER;
         ncplane_putc_yx(hud, 1, x, &ci);
       }
-      if(ncplane_printf_yx(hud, line, 14 - strlen(buf), "%ss", buf) < 0){
+      if(ncplane_printf_yx(hud, line, 14 - ncstrwidth(buf), "%ss", buf) < 0){
         return -1;
       }
       if(ncplane_putstr_yx(hud, line, 16, e->name) < 0){
@@ -612,13 +612,14 @@ int demo_render(struct notcurses* nc){
     if(ncplane_printf_yx(hud, 1, 1, "%d", elems->frames) < 0){
       return -1;
     }
-    char buf[PREFIXCOLUMNS];
-    ncmetric(ns, NANOSECS_IN_SEC, buf, 0, 1000, '\0');
-    for(size_t x = 6 ; x < 14 - strlen(buf) ; ++x){
+    char buf[PREFIXCOLUMNS + 2];
+    ncnmetric(ns, sizeof(buf), NANOSECS_IN_SEC, buf, 0, 1000, '\0');
+    for(int x = 6 ; x < 14 - ncstrwidth(buf) ; ++x){
       nccell ci = CELL_TRIVIAL_INITIALIZER;
       ncplane_putc_yx(hud, 1, x, &ci);
     }
-    if(ncplane_printf_yx(hud, 1, 14 - strlen(buf), "%ss", buf) < 0){
+//fprintf(stderr, "[%s] %zu %d\n", buf, strlen(buf), ncstrwidth(buf));
+    if(ncplane_printf_yx(hud, 1, 14 - ncstrwidth(buf), "%ss", buf) < 0){
       return -1;
     }
     if(ncplane_putstr_yx(hud, 1, 16, elems->name) < 0){